+++ CHANGES.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.0.1
+
+### Fixed
+
+- Module in same project was required in another module, but not add into dependency tree.
+
+- Module in same project was required in another module, but reexport not resolve. Module reexport bundle not pass to another module #3 ([issue](https://github.com/JaneWardSandy/eclipse-pde-partial-idea/issues/3))
+  
+- Manifest bundle reference in same project, linked to wrong module
+
+- Kotlin's property ext access inspection not work #4 ([issue](https://github.com/JaneWardSandy/eclipse-pde-partial-idea/issues/4))
+
+- Kotlin project alert kotlin bundle not required #5 ([issue](https://github.com/JaneWardSandy/eclipse-pde-partial-idea/issues/5))
+
 ## 1.0.0
 
 - Initial project, migrating